+++ noodles-sam/src/header.rs
@@ -48,8 +48,8 @@
 //!
 //! let header = sam::Header::builder()
 //!     .set_header(header::header::Header::default())
-//!     .add_reference_sequence(ReferenceSequence::new("sq0", 8)?)
-//!     .add_reference_sequence(ReferenceSequence::new("sq1", 13)?)
+//!     .add_reference_sequence(ReferenceSequence::new("sq0".parse()?, 8)?)
+//!     .add_reference_sequence(ReferenceSequence::new("sq1".parse()?, 13)?)
 //!     .build();
 //!
 //! assert!(header.header().is_some());
@@ -57,7 +57,7 @@
 //! assert!(header.read_groups().is_empty());
 //! assert!(header.programs().is_empty());
 //! assert!(header.comments().is_empty());
-//! # Ok::<(), sam::header::reference_sequence::NewError>(())
+//! # Ok::<(), Box<dyn std::error::Error>>(())
 //! ```
 
 mod builder;
@@ -169,13 +169,13 @@ impl Header {
     /// use noodles_sam::{self as sam, header::ReferenceSequence};
     ///
     /// let header = sam::Header::builder()
-    ///     .add_reference_sequence(ReferenceSequence::new("sq0", 13)?)
+    ///     .add_reference_sequence(ReferenceSequence::new("sq0".parse()?, 13)?)
     ///     .build();
     ///
     /// let reference_sequences = header.reference_sequences();
     /// assert_eq!(reference_sequences.len(), 1);
     /// assert!(reference_sequences.contains_key("sq0"));
-    /// # Ok::<(), sam::header::reference_sequence::NewError>(())
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
     pub fn reference_sequences(&self) -> &ReferenceSequences {
         &self.reference_sequences
@@ -193,15 +193,17 @@ impl Header {
     /// let mut header = sam::Header::default();
     /// assert!(header.reference_sequences().is_empty());
     ///
+    /// let reference_sequence = ReferenceSequence::new("sq0".parse()?, 13)?;
+    ///
     /// header.reference_sequences_mut().insert(
-    ///     String::from("sq0"),
-    ///     ReferenceSequence::new("sq0", 13)?,
+    ///     reference_sequence.name().to_string(),
+    ///     reference_sequence,
     /// );
     ///
     /// let reference_sequences = header.reference_sequences();
     /// assert_eq!(reference_sequences.len(), 1);
     /// assert!(reference_sequences.contains_key("sq0"));
-    /// # Ok::<(), sam::header::reference_sequence::NewError>(())
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
     pub fn reference_sequences_mut(&mut self) -> &mut ReferenceSequences {
         &mut self.reference_sequences
@@ -378,8 +380,8 @@ impl fmt::Display for Header {
     ///
     /// let header = sam::Header::builder()
     ///     .set_header(header::header::Header::new(header::header::Version::new(1, 6)))
-    ///     .add_reference_sequence(ReferenceSequence::new("sq0", 8)?)
-    ///     .add_reference_sequence(ReferenceSequence::new("sq1", 13)?)
+    ///     .add_reference_sequence(ReferenceSequence::new("sq0".parse()?, 8)?)
+    ///     .add_reference_sequence(ReferenceSequence::new("sq1".parse()?, 13)?)
     ///     .build();
     ///
     /// let expected = "\
@@ -389,7 +391,7 @@ impl fmt::Display for Header {
     /// ";
     ///
     /// assert_eq!(header.to_string(), expected);
-    /// # Ok::<(), header::reference_sequence::NewError>(())
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         if let Some(header) = self.header() {
@@ -451,11 +453,11 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_fmt() -> Result<(), reference_sequence::NewError> {
+    fn test_fmt() -> Result<(), Box<dyn std::error::Error>> {
         let header = Header::builder()
             .set_header(header::Header::new(header::Version::new(1, 6)))
-            .add_reference_sequence(ReferenceSequence::new("sq0", 8)?)
-            .add_reference_sequence(ReferenceSequence::new("sq1", 13)?)
+            .add_reference_sequence(ReferenceSequence::new("sq0".parse()?, 8)?)
+            .add_reference_sequence(ReferenceSequence::new("sq1".parse()?, 13)?)
             .add_read_group(ReadGroup::new("rg0"))
             .add_read_group(ReadGroup::new("rg1"))
             .add_program(Program::new("pg0"))