+++ noodles-sam/src/header/reference_sequence/name.rs
@@ -1,0 +1,6 @@
+//! SAM header reference sequence name.
+
+pub use crate::record::reference_sequence_name::ParseError;
+
+/// A reference sequence name.
+pub type Name = crate::record::ReferenceSequenceName;